requirements.txt
Django>=1.8,<1.9
django-bootstrap3==6.1.0
django-filter==0.10.0
djangorestframework==3.1.3
django-rest-swagger==0.3.4
